中央側 API を /platform 空間へ統合し授権店舗一覧を /platform/stores/me へ移設する (#415)
